+++ versions/weekly/flake.nix
@@ -2,7 +2,7 @@
   inputs =
     {
       holochain = {
-        url = "github:holochain/holochain/holochain-0.4.0-dev.28";
+        url = "github:holochain/holochain/holochain-0.5.0-dev.0";
         flake = false;
       };